add useRegisteredUser HOC
