+++ packages/commerce-sdk-react/src/test-utils.tsx
@@ -8,9 +8,11 @@
 import {render, RenderOptions} from '@testing-library/react'
 import jwt from 'jsonwebtoken'
 import nock from 'nock'
-import React from 'react'
+import React, {useEffect} from 'react'
 import CommerceApiProvider, {CommerceApiProviderProps} from './provider'
-import {QueryClient, QueryClientProvider} from '@tanstack/react-query'
+import {QueryClient, QueryClientProvider, useQueryClient} from '@tanstack/react-query'
+import {ShopperLoginHelpers, useShopperLoginHelper} from './hooks'
+import exp from 'constants'
 
 export const TEST_CONFIG = {
     proxy: 'http://localhost:3000/mobify/proxy/api',
@@ -38,6 +40,27 @@ const TestProviders = (props: {
             </CommerceApiProvider>
         </QueryClientProvider>
     )
+}
+interface Props {
+    comp: React.ElementType // 👈️ type it as React.ElementType
+}
+
+// TODO: fix the type
+export const withRegisteredUser = (Component: React.ComponentType<Props>) => (props) => {
+    const queryClient = useQueryClient()
+    const loginRegisteredUser = useShopperLoginHelper(ShopperLoginHelpers.LoginRegisteredUserB2C)
+    useEffect(() => {
+        // TODO: Before each clear React Query cache
+        queryClient.removeQueries([{entity: 'customer'}])
+
+        // how to check it a user is already registered here before logging them in
+        loginRegisteredUser.mutate({
+            username: 'kobe@test.com',
+            password: 'Test1234!'
+        })
+    }, [])
+
+    return <Component {...props} loginRegisteredUser={loginRegisteredUser} />
 }
 
 /**
